pyproject.toml
[project]
name = "Cold Mail Generator"
version = "0.1.0"
description = ""
authors = [
    {name = "disha",email = "dishagopani05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-groq (>=0.3.2,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "chromadb (>=1.0.4,<2.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "streamlit (>=1.44.1,<2.0.0)"
    ]

[tool.poetry]
packages = [{include = "llama_3", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
